% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dire_bonjour.R
\name{dire_bonjour}
\alias{dire_bonjour}
\title{Title}
\usage{
dire_bonjour(prenom = "toi")
}
\arguments{
\item{prenom}{Un prénom à spécifier}
}
\value{
la fonction renvoie la concaténation du mot Bonjour et du prénom en argument
}
\description{
Title
}
\examples{
dire_bonjour(prenom='Thierry')
}
